+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# next
+* Added `Data.Functor.Invariant.Generic` module, which exports a
+  `genericInvmap` function on GHC 7.2 or later
+* Make `Proxy` and `Tagged` instances poly-kinded
+
 # 0.2
 * Support deriving `Invariant` and `Invariant2` instances with Template Haskell
 * Added `invmapFunctor`, `invmapContravariant`, `invmap2Bifunctor`, and